move vertical-work specific code from insert-char.ts to vertical-work.ts
